Rewrote the model editor to use a base class instead of an interface so code can be shared between Brawlbox and Ikarus. Also includes SCN0 bug fixes, minor REFF work, and better MDL0 billboard bone support

+++ BrawlLib/System/Windows/Controls/SCN0LightEditControl.cs
@@ -238,7 +238,7 @@ namespace System.Windows.Forms
             if (index >= 0)
             {
                 LightAnimationFrame f = (LightAnimationFrame)listKeyframes.SelectedItem;
-                numFrame.Value = f.Index + 1;
+                numFrame.Value = (decimal)(f.Index + 1);
             }
         }
